Configured controllers, exception handling

+++ src/main/java/com/userservice/service/CardInfoServiceImpl.java
@@ -3,6 +3,8 @@ package com.userservice.service;
 import com.userservice.dto.CardInfoDTO;
 import com.userservice.entity.CardInfo;
 import com.userservice.entity.User;
+import com.userservice.exception.CardInfoAlreadyExistsException;
+import com.userservice.exception.CardInfoNotFoundException;
 import com.userservice.mapper.CardInfoMapper;
 import com.userservice.repository.CardInfoRepository;
 import lombok.RequiredArgsConstructor;
@@ -22,36 +24,51 @@ public class CardInfoServiceImpl implements CardInfoService {
     private final CardInfoMapper cardInfoMapper;
 
     @Override
-    public void createCardInfo(CardInfoDTO cardInfoDTO) {
+    public CardInfoDTO createCardInfo(CardInfoDTO cardInfoDTO) throws CardInfoAlreadyExistsException {
+        if(cardInfoRepository.existsByNumberHolderExpirationDate(cardInfoDTO.getNumber(),
+                cardInfoDTO.getHolder(), cardInfoDTO.getExpirationDate())) {
+            throw new CardInfoAlreadyExistsException("Card with these credentials already exists");
+        }
         CardInfo cardInfo = cardInfoMapper.toCardInfo(cardInfoDTO);
-        cardInfoRepository.createCardInfo(cardInfo);
+        return cardInfoMapper.toCardInfoDTO(cardInfoRepository.createCardInfo(cardInfo));
     }
 
     @Override
-    public CardInfoDTO getCardInfo(Long id) {
+    public CardInfoDTO getCardInfo(Long id) throws CardInfoNotFoundException {
         Optional<CardInfo> cardInfo = cardInfoRepository.getCardInfoById(id);
         if (cardInfo.isPresent()) {
             return cardInfoMapper.toCardInfoDTO(cardInfo.get());
         }
-        return null;
+        else{
+            throw new CardInfoNotFoundException("Card with id: " + id + " not found");
+        }
     }
 
     @Override
-    public Page<CardInfoDTO> getAllCardsInfo(Pageable pageable) {
+    public Page<CardInfoDTO> getAllCardsInfo(Pageable pageable) throws CardInfoNotFoundException {
         Page<CardInfo> cardsInfo = cardInfoRepository.getAllCardsInfo(pageable);
+        if(!cardsInfo.hasContent()) {
+            throw new CardInfoNotFoundException("There are not any cards");
+        }
         return cardInfoMapper.toCardInfoDTOPage(cardsInfo);
     }
 
     @Override
     @Transactional
-    public void updateCardInfoById(Long id, CardInfoDTO cardInfoDTO) {
+    public CardInfoDTO updateCardInfoById(Long id, CardInfoDTO cardInfoDTO) throws CardInfoNotFoundException {
+        if(!cardInfoRepository.existsById(id)){
+            throw new CardInfoNotFoundException("Card with id: " + id + " not found for updating");
+        }
         CardInfo cardInfo = cardInfoMapper.toCardInfo(cardInfoDTO);
-        cardInfoRepository.updateCardInfoById(id, cardInfo);
+        return cardInfoMapper.toCardInfoDTO(cardInfoRepository.updateCardInfoById(id, cardInfo));
     }
 
     @Override
     @Transactional
-    public void deleteCardInfo(Long id) {
+    public void deleteCardInfo(Long id) throws CardInfoNotFoundException {
+        if(!cardInfoRepository.existsById(id)){
+            throw new CardInfoNotFoundException("Card with id: " + id + " not found for deleting");
+        }
         cardInfoRepository.deleteCardInfoById(id);
     }
 }
