+++ src/main/java/com/userservice/service/UserServiceImpl.java
@@ -2,6 +2,9 @@ package com.userservice.service;
 
 import com.userservice.dto.UserDTO;
 import com.userservice.entity.User;
+import com.userservice.exception.CardInfoNotFoundException;
+import com.userservice.exception.UserAlreadyExistsException;
+import com.userservice.exception.UserNotFoundException;
 import com.userservice.mapper.UserMapper;
 import com.userservice.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -21,52 +24,67 @@ public class UserServiceImpl implements UserService {
     private final UserMapper userMapper;
 
     @Override
-    public void createUser(UserDTO dto) {
+    public UserDTO createUser(UserDTO dto) throws UserAlreadyExistsException {
+        if(userRepository.existsByEmail(dto.getEmail())){
+            throw new UserAlreadyExistsException("User with email: " + " already exists");
+        }
         User u = userMapper.toUser(dto);
-        userRepository.createUser(u);
+        return userMapper.toUserDTO(userRepository.createUser(u).get());
     }
 
     @Override
-    public UserDTO getUserById(Long id) {
+    public UserDTO getUserById(Long id) throws UserNotFoundException {
         Optional<User> u = userRepository.getUserById(id);
 
         if(u.isPresent()) {
             return userMapper.toUserDTO(u.get());
         }
-        return null;
+        else{
+            throw new UserNotFoundException("User with id: " + id + " not found");
+        }
 
     }
 
     @Override
-    public Page<UserDTO> getAllUsers(Pageable pageable) {
+    public Page<UserDTO> getAllUsers(Pageable pageable) throws CardInfoNotFoundException {
 
         Page<User> users = userRepository.getAllUsers(pageable);
-        Page<UserDTO> dtos = userMapper.toUserDTOPage(users);
-        return dtos;
+        if(!users.hasContent()) {
+            throw new CardInfoNotFoundException("There are not any users");
+        }
+        return userMapper.toUserDTOPage(users);
 
     }
 
     @Override
-    public UserDTO getUserByEmail(String email) {
+    public UserDTO getUserByEmail(String email) throws UserNotFoundException {
 
         Optional<User> u = userRepository.getUserByEmail(email);
         if(u.isPresent()) {
             return userMapper.toUserDTO(u.get());
         }
-        return null;
+        else{
+            throw new UserNotFoundException("User with email: " + email + " not found");
+        }
 
     }
 
     @Override
     @Transactional
-    public void updateUser(Long id, UserDTO dto) {
+    public UserDTO updateUser(Long id, UserDTO dto) throws UserNotFoundException {
+        if(!userRepository.existsById(id)){
+            throw new UserNotFoundException("User with id: " + id + " not found for updating");
+        }
         User u = userMapper.toUser(dto);
-        userRepository.updateUserById(id, u);
+        return userMapper.toUserDTO( userRepository.updateUserById(id, u));
     }
 
     @Override
     @Transactional
-    public void deleteUser(Long id) {
+    public void deleteUser(Long id) throws UserNotFoundException {
+        if(!userRepository.existsById(id)){
+            throw new UserNotFoundException("User with id: " + id + " not found for deleting");
+        }
         userRepository.deleteUserById(id);
     }
 }
